inference-backend/crowd-analytics/requirements.txt
certifi
charset-normalizer
contourpy
cycler
filelock
fonttools
fsspec
idna
Jinja2
kiwisolver
MarkupSafe
matplotlib
mpmath
networkx
numpy
nvidia-cublas-cu12; platform_system == "Linux"
nvidia-cuda-cupti-cu12; platform_system == "Linux"
nvidia-cuda-nvrtc-cu12; platform_system == "Linux"
nvidia-cuda-runtime-cu12; platform_system == "Linux"
nvidia-cudnn-cu12; platform_system == "Linux"
nvidia-cufft-cu12; platform_system == "Linux"
nvidia-cufile-cu12; platform_system == "Linux"
nvidia-curand-cu12; platform_system == "Linux"
nvidia-cusolver-cu12; platform_system == "Linux"
nvidia-cusparse-cu12; platform_system == "Linux"
nvidia-cusparselt-cu12; platform_system == "Linux"
nvidia-nccl-cu12; platform_system == "Linux"
nvidia-nvjitlink-cu12; platform_system == "Linux"
nvidia-nvtx-cu12; platform_system == "Linux"
opencv-python
packaging
pandas
pillow
psutil
py-cpuinfo
pyparsing
python-dateutil
pytz
PyYAML
requests
scipy
six
sympy
torch
torchvision
tqdm
triton; platform_system == "Linux"
typing_extensions
tzdata
ultralytics
ultralytics-thop
urllib3
